refactor: main UI 패키지 변경
